Moved static variables into game class

+++ notePlayer.hpp
@@ -1,5 +1,4 @@
 #ifndef ARDUINO_H
-
 #define ARDUINO_H
 #include "Arduino.h"
 #endif
@@ -11,7 +10,7 @@ class NotePlayer{
     // Buzzer Pin
     uint8_t tonePin = 0;
     // Note length in ms
-    uint8_t toneLength = 500;
+    int toneLength = 500;
     // Note sequence feedback
     static uint16_t badNoteSequence[];
     static uint16_t goodNoteSequence[];
